+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to Rust's notion of
   - `FieldExt::{from_bytes, read, to_bytes, write}`
     (use `ff::PrimeField::{from_repr, to_repr}` instead).
   - `FieldExt::rand` (use `ff::Field::random` instead).
+  - `CurveAffine::{read, write}`
+    (use `group::GroupEncoding::{from_bytes, to_bytes}` instead).
 
 ## [0.2.1] - 2021-09-17
 ### Changed